backend/requirements.txt
flask
pytest
pytest-mock
requests
cryptography
PyJWT